backend/requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
fire==0.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
Pygments==2.19.2
PyYAML==6.0.2
rav==0.0.9
rich==14.0.0
sqlparse==0.5.3
termcolor==3.1.0
tzdata==2025.2
